+++ common/src/main/kotlin/com/github/jcornaz/collekt/Interfaces.kt
@@ -5,18 +5,20 @@ package com.github.jcornaz.collekt
  *
  * All implementers promise that the collection won't ever change after instance creation
  */
-interface ImmutableCollection<out E> : Collection<E>
+public interface ImmutableCollection<out E> : Collection<E>
 
 /**
  * An immutable [List]
  *
  * All implementers promise that the list won't ever change after instance creation
  */
-interface ImmutableList<out E> : ImmutableCollection<E>, List<E>
+public interface ImmutableList<out E> : ImmutableCollection<E>, List<E> {
+    override fun subList(fromIndex: Int, toIndex: Int): ImmutableList<E>
+}
 
 /**
  * An immutable [Set]
  *
  * All implementers promise that the list won't ever change after instance creation
  */
-interface ImmutableSet<out E> : ImmutableCollection<E>, Set<E>
+public interface ImmutableSet<out E> : ImmutableCollection<E>, Set<E>