+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 0.0.1-SNAPSHOT (Unreleased)
 * Add Immutable collection interfaces: `ImmutableCollection`, `ImmutableList`, `ImmutableSet`
 * add `emptyImmutableList()` and `emptyImmutableSet()`
+* Add `immutableListOf()`, `Iterable.toImmutableList()`, `Sequence.toImmutableList()` and `List.asImmutableList()`